fix: cleanup code and remove unsused endpoints
