+++ tsconfig.json
@@ -15,8 +15,7 @@
     "jsx": "preserve",
     "strictPropertyInitialization": false,
     "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "downlevelIteration": true
+    "emitDecoratorMetadata": true
   },
   "include": ["next-env.d.ts", "src/**/*.ts", "src/**/*.tsx"],
   "exclude": ["node_modules"]
